Restructuring and adding icons
