+++ .gitignore
@@ -1,0 +1,5 @@
+.DS_Store
+private/
+/locale
+node_modules/
+*.log
