+++ .npmignore
@@ -1,0 +1,29 @@
+.DS_Store
+.settings/
+.sass-cache/
+.project
+.idea
+npm-debug.log*
+yarn-error.log
+yarn.lock
+style/.sass-cache/
+stats.json
+
+img/
+images/
+private/
+vendor/
+coverage/
+node_modules/
+bower_components/
+grapesjs-*.tgz
+_index.html
+index.html
+docs
+.github
+test
+.editorconfig
+.eslintrc
+.travis.yml
+*.md
+webpack.config.js
