+++ .travis.yml
@@ -1,0 +1,4 @@
+language: node_js
+node_js:
+  - "8"
+  - "10"
