+++ README.md
@@ -7,6 +7,8 @@ Plugin for GrapesJS that adds table block.
 |Option|Description|Default|
 |-|-|-
 |`containerId`|Container to wich will be binded jQuery Event handlers to control adding new table and handler for adding and removing rows and columns.|`document`|
+|`tblResizable`|Should the table have resize handles|`true`|
+|`cellsResizable`|Should cells within the table have resize handles|`true`|
 
 ## Usage
