+++ package.json
@@ -1,0 +1,32 @@
+{
+  "name": "grapesjs-blocks-table",
+  "version": "0.0.31",
+  "description": "grapesjs-blocks-table",
+  "main": "dist/grapesjs-blocks-table.min.js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/VojtechLanka/grapesjs-blocks-table.git"
+  },
+  "scripts": {
+    "start": "grapesjs-cli serve",
+    "build": "grapesjs-cli build",
+    "bump": "npm version patch -m 'Bump v%s'"
+  },
+  "keywords": [
+    "grapesjs",
+    "plugin",
+    "table",
+    "grapesjs-table"
+  ],
+  "devDependencies": {
+    "grapesjs-cli": "^1.0.13"
+  },
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/VojtechLanka/grapesjs-blocks-table/issues"
+  },
+  "homepage": "https://github.com/VojtechLanka/grapesjs-blocks-table#readme",
+  "dependencies": {
+    "jquery": "^3.6.0"
+  }
+}
