+++ src/index.js
@@ -1,0 +1,21 @@
+import onLoad from './onLoad'
+import components from './components';
+import blocks from './blocks';
+import commands from './commands';
+
+export default (editor, opts = {}) => {
+  const options = {
+    ...{
+      tblResizable: true,
+      cellsResizable:true,
+    },
+    ...opts,
+  };
+
+  onLoad();
+  
+  components(editor, options)
+  blocks(editor, options)
+  commands(editor, options)
+};
+
